VC-STOCK-BETA/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
reportlab
Pillow
fpdf
python-barcode
